Separação por modulos e alguns ajustes
