+++ compiladores/python/interpretador.py
@@ -1,15 +1,14 @@
-from parser import Parser
-from lexer import Lexer
+from compiladores.python.modules.Parser import Parser
+from compiladores.python.modules.AnaliseLexica import AnaliseLexica
 
 class Interpretador:
     def __init__(self):
-        self.lexer = Lexer()
+        self.analiseLexica = AnaliseLexica()
         self.parser = Parser()
         self.run()
 
     def run(self):
-        code = 'A = B + C; C = B; A = B - C;'
-        tokens = self.lexer.analise_lexica(code, Lexer().tokens_da_linguagem)
+        tokens = self.analiseLexica.analise_lexica('A = B + C; C = B; A = B - C;')
         print(tokens)
         programa = self.parser.analise_sintatica(tokens)
         print (programa)
